requirements.txt
# FastAPI Core
fastapi>=0.104.0,<0.105.0
uvicorn[standard]>=0.24.0,<0.25.0
python-multipart>=0.0.6,<0.1.0

# Procesamiento de imágenes y PDFs
opencv-python>=4.8.0,<5.0.0
numpy>=1.24.0,<2.0.0
Pillow>=10.0.0,<11.0.0
pypdfium2>=4.26.0,<5.0.0

# OCR engines
easyocr>=1.7.0,<2.0.0
pytesseract>=0.3.10,<0.4.0

# Utilities
python-dotenv>=1.0.0,<2.0.0
requests>=2.31.0,<3.0.0
